pyproject.toml
[project]
name = "micro-videos-python"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "guilherme", email = "vinespedro55@gmail.com"},
]
dependencies = [
    "pytest>=8.1.1",
    "flake8>=7.0.0",
    "Django>=5.0.3",
    "djangorestframework>=3.15.0",
    "pytest-django>=4.8.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
pytest = [
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
